Move AuxLosses to common
