+++ habitat_baselines/models/rcm/vln_rcm_policy.py
@@ -12,6 +12,7 @@ import torch.nn.functional as F
 from gym import Space
 
 from habitat import Config
+from habitat_baselines.common.aux_losses import AuxLosses
 from habitat_baselines.common.utils import CategoricalNet, Flatten
 from habitat_baselines.models.instruction_encoder import InstructionEncoder
 from habitat_baselines.models.rcm.rcm_state_encoder import RCMStateEncoder
@@ -26,7 +27,6 @@ from habitat_baselines.models.simple_cnn import (
     SimpleRGBCNN,
 )
 from habitat_baselines.rl.ppo.policy import Net, Policy
-from habitat_baselines.rl.aux_losses import AuxLosses
 
 
 class VLNRCMPolicy(Policy):
